[UXE-7010] fix: change query to show invoice details correctly to trial users (#2363)
* fix: change query to show invoice details correctly to trial users

* test: add e2e test to this case

* fix: add a helper to do the date parsing

* test: fix import

* test: remove e2e flaky tests
