+++ src/helpers/get-current-month-start-end.js
@@ -1,0 +1,13 @@
+export const getCurrentMonthStartEnd = () => {
+  const today = new Date()
+  const year = today.getFullYear()
+  const month = today.getMonth()
+
+  const dateInitial = new Date(year, month, 1)
+  const dateFinal = new Date(year, month + 1, 0)
+
+  return {
+    dateInitial: dateInitial.toISOString().split('T')[0],
+    dateFinal: dateFinal.toISOString().split('T')[0]
+  }
+}
