+++ src/services/billing-services/load-current-invoice-service.js
@@ -2,6 +2,7 @@ import { AxiosHttpClientAdapter, parseHttpResponse } from '../axios/AxiosHttpCli
 import { makeBillingBaseUrl } from './make-billing-base-url'
 import { makeAccountingBaseUrl } from './make-accounting-base-url'
 import { formatDateToUSBilling } from '@/helpers/convert-date'
+import { getCurrentMonthStartEnd } from '@/helpers/get-current-month-start-end'
 import { useAccountStore } from '@/stores/account'
 
 export const loadCurrentInvoiceService = async () => {
@@ -52,20 +53,6 @@ const adapt = (httpResponse, accountIsNotRegular) => {
   }
 }
 
-const getCurrentMonthStartEnd = () => {
-  const today = new Date()
-  const year = today.getFullYear()
-  const month = today.getMonth()
-
-  const dateInitial = new Date(year, month, 1)
-  const dateFinal = new Date(year, month + 1, 0)
-
-  return {
-    dateInitial: dateInitial.toISOString().split('T')[0],
-    dateFinal: dateFinal.toISOString().split('T')[0]
-  }
-}
-
 const getQueryByAccountType = (accountIsNotRegular) => {
   const dateRange = getCurrentMonthStartEnd()
 
@@ -90,6 +77,9 @@ const getQueryByAccountType = (accountIsNotRegular) => {
             periodTo,
             invoiceNumber,
             totalValue,
+            currency,
+            temporaryBill
+        }
       }`
     }
   } else {
